django-2-SQL/requirements.txt
asgiref==3.8.1
Django==5.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.3
